// src/app/core/services/http-client/model/request-options.ts
import {HttpHeaders, HttpParams} from '@angular/common/http';

export interface RequestOptions {
  body?: any;
  headers?: HttpHeaders | { [header: string]: string | string[]; };
  observe?: 'body';
  params?: HttpParams | { [param: string]: string | string[]; };
  reportProgress?: boolean;
  responseType?: 'arraybuffer' | 'blob' | 'json' | 'text';
  withCredentials?: boolean;
}